refactor: Adapt activity image handling to support multiple attachments
The `beforeImage` and `afterImage` fields in the `Activity` type have been changed from single `Attachment` objects to arrays of `Attachment` objects. This allows for multiple images to be associated with the "before" and "after" states of an activity.

Mock data, form components, and the Kanban card display have been updated to reflect this change, supporting the display and upload of multiple images.